Create a JS function to play the game from the console + minor fixes
The function game() allows the user to start the game from the console.

There are numOfRounds = 5 rounds and the player/computer scores are updated
on each round depending on who wins.

User choices are retrieved through a prompt.
Results messages are shown to the users directly in the console.

Some simple functions have been written to compose these results messages.

Also, the possibility to show these results messages in an alert instead of
on the console is achieved by starting the game with game(alert).
To do this, the output function is specified as a generic printFcn(), which is
an optional parameter to game(), defaulting to console.log().