Add terminal interface (jQuery terminal) to play the game directly on the webpage
jQuery terminal is a plugin that allows to emulate a terminal
directly in a webpage.
See https://terminal.jcubic.pl/api_reference.php

This way it is possible to play the first version of the game
without having to open the browser console, i.e., also for
mobile users, at least until the final version with graphical
interface is completed.

In this commit, you can't play the game in the terminal just yet.
Only a simple hello world command is added, to make sure
everything works.