Add player and computer animations
Add an animation for the player and the computer.

A Lottie file has been used, which allows to control which parts of
the animations to show depending on the phase in which the game is
on, like 'waiting for the player to choose its move' or 'showing the
players moves' and so on.