Add animations to the moves buttons + minor fixing
Representative animations have been included in the buttons
representing the possible moves, to make it more immediate the
meaning of the moves.

The layout has also been slightly modified.

A minor issue is still present (the layout moves of few pixels
when switching from the 'choose a move' to the 'show round
outcome' phases), which will be addressed in the future.